IDC-PROJECT/requirements.txt
groq==0.10.0
streamlit==1.38.0
torch 2.4.1
torchvision 0.19.1
pillow 10.4.0